Use different AssetDatabase API call to make it work in 2019.4
The method `AssetDatabase.GUIDFromAssetPath` is introduced in Unity 2020.2 so
the package had a compile error in versions before that (I have only tested it
on 2019.4).

The amount of changes needed for this fix were minimal as the code was already
working with the string value of the GUID.

However, there might be a notable change in behavior - according to the
documentation, `AssetPathToGUID` will return assets that were deleted during the
current editor session (you'd have to restart the Unity Editor to make it not
include those). The docs for `GUIDFromAssetPath` don't mention that behavior.
From 2021.1 they've added an optional argument to `AssetPathToGUID` that allows
you to get around this, but I didn't use it here because it doesn't exist in
earlier versions. If needed, this could be improved in a future commit by using
`#if UNITY_2021_1_OR_NEWER`.